Fix casting with pow - can be returned float in some cases when result > PHP_INT_MAX

+++ tests/Helpers/Generator/ProvidesRandomizedValues.php
@@ -43,7 +43,9 @@ trait ProvidesRandomizedValues
 
     private function powerOfTen(int $scale): int
     {
-        return 10 ** $scale;
+        /* @noinspection UnnecessaryCastingInspection */
+        /* @noinspection PhpCastIsUnnecessaryInspection */
+        return (int) (10 ** $scale);
     }
 
     private function safeUnitsUpperBound(int $scale, int $maxFactor = 9): int
